fpga: dfl: VFIO mdev support for DFL devices
This patch is to provide VFIO support to DFL devices. When bound this
driver to a dfl device, it registers a set of callbacks to the mediated
device framework to enable mediated device creation. When a mediated
device is created by user via mdev interfaces, it will be exposed to
user as a VFIO platform device.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>
Reviewed-by: Wu Hao <hao.wu@intel.com>
----
v2: delete the iommu_device part, cause now we assume none of the dfl
    device has dma capability.
v3: split the driver_override stuff to a separate patch.
    Change the driver name, "vfio-dfl" -> "vfio-mdev-dfl"