fpga: dfl: fme: enable secure updates
Extend security manager to include an update/filename
sysfs file that can be used to initiate a security update.
The filename of a secure update file (bmc image, bmc
firmware, static region image, etc) can be written to
this sysfs entry to cause a secure update to occur.

The write will return immediately, and the update will
begin in the context of a kernel worker thread. Note
that this tool utilizes the request_firmware framework.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>