fpga: dfl: add support for switching the retimer FEC mode
The N3000 Nios private feature provides the one-time configuration for
the retime FEC mode. The Nios firmware is responsible for the actual
hardware operation. It gives users a chance (only one chance) to set the
retimers to a different FEC mode.

The Nios firmware has no nvmem to store the required FEC mode info, so it
needs the user input everytime on board power up. So this patch
introduced a module parameter for the user input.

However, the community rejects this module parameter solution. Firstly
the module parameter can not support different configuration for
multiple boards. Secondly the maintainer thinks this configuration
should not be cared by linux user, the bootloader (thus the on board
firmware) should be changed to handle the configuration.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>