fpga: dfl: implement dfl_dev_get_base_dev()
The n5010-hssi driver added in one of the upcoming commits needs a way to
look up its related n5010-phy driver. For this it needs a way to compare
the base device for the two driver instances.

Support this use case by implementing dfl_dev_get_base_dev() in the dfl
bus.

Signed-off-by: Martin Hundebøll <mhu@silicom.dk>