mfd: intel-m10-bmc: Handle d5005 fw handshake registers
Some of the MAX10 BMC registers can become stale if they are not
regularly updated by the BMC firmware. The prime examples are the
telemetry registers which require frequent updates. During a
secure update of the FPGA or BMC images, we cause reads to these
registers to faile with -EBUSY, because the BMC firmware cannot
keep them up to date during a secure update.

Currently, the driver only supports the n3000 telemetry register
range. This change adds a specific range for the d5005 telemetry
register range.

- Russ