fpga: m10bmc-sec: create max10 bmc secure update driver
Create a sub driver for the FPGA Card BMC in order to support secure
updates.  This sub-driver will invoke an instance of the FPGA Image Load
class driver for the image load portion of the update.

This patch creates the MAX10 BMC Secure Update driver and provides sysfs
files for displaying the current root entry hashes for the FPGA static
region, the FPGA PR region, and the MAX10 BMC.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>
Reviewed-by: Tom Rix <trix@redhat.com>