fpga: m10bmc-sec: support power on image
This patch supports the ability to set a default power-on FPGA image.
Additionally, one can select the fall-back image sequence. Two sysfs
files are added to the Intel MAX10 BMC Secure driver:

available_fpga_images (RO)
power-on-image        (WO)

The expected string to be read from available_fpga_images is:

        fpga_factory fpga_user1 fpga_user2

A string to be written to power-on-image is expected to be one of
the following:

        fpga_user1
        fpga_user2
        fpga_factory
        fpga_factory fpga_user1
        fpga_factory fpga_user2

Signed-off-by: Russ Weight <russell.h.weight@intel.com>
Signed-off-by: Tianfei Zhang <tianfei.zhang@intel.com>