configs: DFL driver config options
Having separate config files means having a lot of duplicate options
listed in both files. It also means that both config files need to be
used in order to build all of the DFL support into a single kernel.

This change merges the configs into a single list of config options and
renames the readme.txt file to README.

The n3000_d5005_defconfig and n5010_defconfig files are now deprecated
and will be removed soon.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>