fpga: m10bmc-sec: add sysfs to load bmc images
Add the available_images and image_load syfs files. The available_images
file returns a space separated list of key words that may be written
into the image_load file. These keywords decribe an FPGA, BMC, or
firmware image in FLASH or EEPROM storage that may be loaded.

The image_load syfs file may be written with a key word to trigger a
reload of an FPGA, BMC, or firmware image from FLASH or EEPROM.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>