fpga: m10bmc-sec: Add trigger to read SDM key hashes
Trigger the reading of the SDM key hashes to BMC registers during driver
initialization and check for completion of the read before completing
a sysfs read for the same data. The SDM trigger is executed in a
worker thread and flush_work() is used to guarantee that the work
has completed before responding to a sysfs read of the data.

Signed-off-by: Tianfei Zhang <tianfei.zhang@intel.com>
Signed-off-by: Russ Weight <russell.h.weight@intel.com>