fpga: m10bmc-sec: Fix load of sdm provisioning data
The SDM key provisioning information is loaded to BMC registers during
driver initialization so that they can be read by fpgainfo security.
The driver has to trigger this request. Currently the driver checks to
see if it is already done before doing it. This means that loading the
driver alone is not enough to cause a new load of the data. Furthermore,
in testing, even after a BMC reset, the "done" status is there during
driver initialization and so the provisioning data is never actually
loaded.

Change the driver to unconditionally trigger a load of the SDM
provisioning data when it loads. This will also facilitate the provisioning
process, allowing new keys to be viewed in fpgainfo security after an rsu
command causes a reload of the driver.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>