fpga: m10bmc-sec: Fix error messages for SCM CMD
If an error happens when polling for the SDM command to complete, the
status field of the SDM command register is printed. In this case, it
would be better to print the error field. It is also helpful to
differentiate between an error from regmap_read_poll_timeout() vs an
error code in the SDM command register.

In addition to making these changes, the wording of the related
error messages is modified and where an error return (ret) is
already available, return that instead of specifically returning
-EIO.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>