mfd: intel-m10-bmc: Change staging size to a variable
The size of the staging area in FLASH for FPGA updates is dependent on the
size of the FPGA. Currently, the staging size is defined as a constant.
Larger FPGAs are coming soon and it will soon be necessary to support
different sizes for the staging area. Add a new staging_size member to the
csr_map structure to support a variable staging size.

The secure update driver does a sanity-check of the image size in
comparison to the size of the staging area in FLASH. Change the
staging size reference to a variable instead of a constant in order
to more readily support future, larger FPGAs.

Co-developed-by: Russ Weight <russell.h.weight@intel.com>
Signed-off-by: Russ Weight <russell.h.weight@intel.com>
Signed-off-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com>