net: phy: qsfp-mem: Fix platform driver name
Change the driver name from "intel,qsfp-mem" to "qsfp-mem". This name shows
up in the sysfs path names where a comma within the name is non-standard.
Also, using a consistent name of "qsfp-mem" for both the dfl and platform
drivers will create a consistent pathnames for sysfs files between the two
versions of the driver.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>