hwmon: intel-m10-bmc: constify pointers to hwmon_channel_info
Statically allocated array of pointed to hwmon_channel_info can be made
const for safety.

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>

(cherry picked from commit e374c4100b6154f80aaf12af39c67a942b519a87)
For the 6.1-lts kernel, it was necessary to cast the value that is
assigned to hw->chip.info, as the upstream changes are not all present
for the 6.1-lts kernel at this time.

Signed-off-by: Signed-off-by: Russ Weight <russell.h.weight@intel.com>