fpga: dfl: Add platform driver for DFL device tree
In order to support the integration of the Device Feature List (DFL) into
non-PCIe FPGA System-on-Chip (SoC) solutions, the Device Tree Node is used
to define the DFL location within the Memory-Mapped I/O (MMIO) space.
To process the device tree node, a driver module has been implemented and
utilizes the information within the Linux DFL framework's (dfl.ko) module
for the purpose of enumerating and managing the FPGA Features.

Signed-off-by: Basheer Ahmed Muddebihal <basheer.ahmed.muddebihal@linux.intel.com>